pybackend/requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-admin-material-dashboard==1.0.10
django-jazzmin==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
msgpack==1.0.8
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python3-openid==3.2.0
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
corsheaders==3.7.0
